FIX TypeError exception on recent OpenCV versions
OpenCV v3.4+ raises a TypeError when a non-numpy array is fed to cv2.mean

This is fixed for the explicit array feed-in on line 384 in detectors.py.

This may occur also on the locations where it is assumed that a frame
returned by a cv2 capture object is always a numpy array, however this
is a safe assumption.