Support selection of scenes for `save-images` and `split-video`
Add a new `-r` / `--scene-range` command-line option to `scenedetect`
that allows users to specify the scene indxes they'd like to save images
and/or videos from.  This option should be a comma-separated list of one
or more integers and/or ranges.  Ranges are specified in a syntax
similar to python's list slice notation, e.g. `start:[stop:[step]]`.

Scene detection will stop once the number of scenes to be selected has
been detected.  (If any of the indexes is specified as a negative offset
from the end, then all scenes must be detected, and no such
short-circuiting is possible.)