Add `save-images` option for ignoring a number of frames at start/end.
Add `--image-frame-margin` option, which specifies a number of frames
to ignore at the beginning and end of scenes when generating images.

This is useful if scene cuts may have frames from the previous/next
scene in them for some reason, or if there are some transition effects
or camera movement between scenes.