[feature] Integrate edge detector
Based off https://github.com/wjs018/QuantitativeEditing (thanks @wjs018).

Works well with referenced material in #35 but many outstanding TODOs
remain in regards to integration with the rest of the application.

Remove support for loading data in ContentDetector from statsfile as this
functionality is now deprecated (stats will always be re-calculated). The
same functionality can be achieved if/when required by implementing another
ContentDetector which reads stats from a file.