[cli] Move output command state out of CliContext
To simplify things, all output commands can be treated similarily
as they all act on the result of the processing pipeline. This
allows new commands to be added without needing to explicitly define
their values in CliContext, and makes the values that do remain there
much more meaningful.

It also allows commands to be specified multiple times gracefully
and with different options, so for example `save-images` can now be
run twice with different encoding parameters.