[detector] Add new detector interface and implment it for ContentDetector
Fixing VFR videos requires detectors have knowledge of timestamps, so we should tackle the detector API overhaul now.

This commit introduces the new API as well as a basic implementation for ContentDetector to show what needs to be changed.

The new detector is used when specifing `detect-content -b`.  It does not yet enforce min_scene_len, and may still have bugs.