[timecode] Add new Timecode and Framerate types
The existing FrameTimecode type assumes fixed framerate, and relies on it for calculation of durations.

This splits both out, and adds timing information in addition to frame number.

This will need to be integrated in the VideoStream side of things first, before starting to work through SceneManager and the Detector interface.